Move magnetic moment tests to PBC tools
